chore(build.yml): update Redis image version to redis:7.0.12-alpine
The Redis image version is updated to redis:7.0.12-alpine to use the
latest version of Redis. The Go version is updated to ~1.20.6 to allow
for compatibility with the latest Go version. The environment variable
names for the Redis service are changed to REDIS_SERVICE_HOST and
REDIS_SERVICE_PORT for clarity. The codecov-action version is updated to
v3 for the latest features and improvements.