fix(ci.yml): rename 'build' job to 'call-docker-build' and use a reusable workflow for docker build and push
The 'build' job has been renamed to 'call-docker-build' to better
reflect its purpose. Instead of defining the steps directly in the
workflow file, a reusable workflow located at
'dictyBase/workflows/.github/workflows/docker-build-push.yaml@develop'
is now used. This improves maintainability and reduces duplication of
code. The secrets are inherited from the parent workflow and the
repository and ref are passed as inputs to the reusable workflow.