refactor(server.go): remove checkEndpoints function and related code
The checkEndpoints function and related code were removed because they
were causing unnecessary HTTP requests. These requests were used to
check the status of API endpoints, but this is not necessary as the
status of these endpoints can be assumed to be OK. This change reduces
the complexity of the code and improves performance by eliminating
unnecessary network calls.