feat(repository.go): add SetWithTTL method to Repository interface
The SetWithTTL method is added to the Repository interface to allow
setting a value with a specified Time To Live (TTL). This feature
provides the ability to have keys in the repository that automatically
expire after a certain duration, which can be useful for caching or
other temporary data storage needs.