feat(authentication/client.go): add User function to fetch user details
The User function has been added to the LogtoClient struct in the
authentication package. This function fetches user details from the
server using the user's ID. It first retrieves the token, then makes a
GET request to the server. If the request is successful, it decodes the
JSON response into a UserResp struct and returns it.