feat(client.go): add UserWithEmail function to retrieve user data by email
The userId variable was renamed to userStruct to better reflect its
purpose and usage in the code. This change improves code readability and
maintainability. The new function UserWithEmail was added to provide a
way to retrieve user data using an email address. This function enhances
the functionality of the client by allowing user data retrieval by
email, which can be more convenient in certain use cases.