refactor(content.go): add detailed user information to UpdatedBy method
The unused import
'github.com/dictyBase/go-genproto/dictybaseapis/api/jsonapi' was removed
to clean up the code. The method of retrieving a user was updated from
using an ID to using an email, which is more reliable and user-friendly.
The UpdatedBy method now returns detailed user information, providing a
more comprehensive user profile.