refactor(client.go): convert LogtoClient struct to interface and create authClient struct
The LogtoClient struct is converted to an interface to provide a clear
contract for what a LogtoClient should be able to do. This allows for
better abstraction and easier testing as we can now create mock
implementations of the LogtoClient interface. The actual implementation
of the LogtoClient is now done in the authClient struct.