refactor(auth_client_mock.go): update mock functions to reflect changes in authentication service
The mock functions in the auth_client_mock.go file have been updated to
reflect changes in the authentication service. The AuthServiceClient
mock type has been renamed to LogtoClient, and the mock functions have
been updated to match the new methods in the LogtoClient interface. This
includes new mock functions for AccessToken, AddCustomUserInformation,
CheckUser, CheckUserWithUserName, CreateUser, Permissions, Roles, User,
and UserWithEmail. The changes were necessary to keep the tests
up-to-date with the changes in the authentication service.