feat(.golangci.yml): add depguard rule to deny usage of io/ioutil package
The depguard rule has been added to the golangci-lint configuration to
deny the usage of the io/ioutil package. This is because the io/ioutil
package has been replaced by io and os packages since Go 1.16. This
change ensures that the codebase stays up-to-date with the latest Go
standards and practices.