feat(mutation.generated.go): add UploadFile function to MutationResolver interface
The UploadFile function was added to the MutationResolver interface to
allow for file uploads. This function was then implemented in the
MutationResolver. A new file, upload.generated.go, was created to handle
the logic for file uploads. This includes marshalling and unmarshalling
of the uploaded file, as well as error handling. This change allows
users to upload files through the GraphQL API.