refactor(content.go): move content scope variables to authentication package
The content scope variables (contentCreatorScope, contentEditorScope,
contentDeleteScope) were moved to the authentication package. This
change was made to improve the organization of the code. By grouping
related variables in the same package, we make the code easier to
understand and maintain. It also ensures that these variables are
available to other parts of the application that might need them,
promoting code reuse.