feat(go.mod): add direct dependency on github.com/google/uuid v1.6.0
The direct dependency on github.com/google/uuid v1.6.0 was added to use
UUID functionality in the application. The
google.golang.org/genproto/googleapis/rpc was updated to the latest
version to incorporate the latest features and bug fixes. The indirect
dependency on gopkg.in/ini.v1 v1.67.0 was added to parse INI files,
which is required by some other dependency.