chore(ci): rename job from call-docker-build to call-build-deploy for clarity and update workflow to use composite-deploy.yaml
The job name is updated to better reflect its purpose, which is to build
and deploy the application. The workflow is also changed to use
composite-deploy.yaml, allowing for a more streamlined deployment
process. Additional parameters for the application, project, stack,
dockerfile, and docker_image are included to ensure proper configuration
during the deployment.